refactor(admin): extract Health tab into pages/admin/health.vue
